Fixed bug - Terms using wrong ID
In f12bfc9687ab6a149964c079d89964f138b49c8d, attaching terms to posts was
made to work, but with a bug.  The wrong ID was used for sending the
desired terms to WordPress.  Instead of using the Id of a term, which
represents the term itself, the POST request is using the TermTaxonmyId,
which represents the association of the term to a taxonomy.  As a result,
I was having problems in some sites but not others - the difference being
that in sites where the only terms are categories, the IDs wound up being
the same and everything was fine.  In sites where this was not the case, I
was getting either a 403 fault code - Invalid Term ID, or attaching the
wrong term.