Wrong member declaration type
I noticed sized were being returned as null in my test project. that's because of the wrong member declaration.

I think members should be of type "MediaItemSize" (see change request for typo in that class also)

Also, [XmlRpcMember("post_thumbnail")] on Line 17 is not correct. Rest of members were returning ok, but this one was still returning null. I looked over the XML RPC api for wordpress and it is called "post-thumbnail" (dash instead of underline).

Hope this helps.